+++ src/library/state/constants.rs
@@ -1,45 +1,45 @@
-pub const NODATAVAL: f64 = -9999.0;
+pub const NODATAVAL: f32 = -9999.0;
 
-pub const  MODISSNOWVAL :f64 = 200.0;
-pub const  DFFM_DEFAULT:f64 = 20.0;
+pub const  MODISSNOWVAL :f32 = 200.0;
+pub const  DFFM_DEFAULT:f32 = 20.0;
 
-pub const  MAXRAIN:f64 = 0.1;
-pub const  A1:f64 = 1.0;
-pub const  A2:f64 = 0.555;
-pub const  A3:f64 = 10.6;
-pub const  A4:f64 = 0.5022;
-pub const  A5:f64 = 0.0133;
-pub const  A6:f64 = 0.000343;
-pub const  A7:f64 = 0.00722;
+pub const  MAXRAIN:f32 = 0.1;
+pub const  A1:f32 = 1.0;
+pub const  A2:f32 = 0.555;
+pub const  A3:f32 = 10.6;
+pub const  A4:f32 = 0.5022;
+pub const  A5:f32 = 0.0133;
+pub const  A6:f32 = 0.000343;
+pub const  A7:f32 = 0.00722;
 	
-pub const  B1:f64 = 3.0;
-pub const  B2:f64 = 0.60;
-pub const  B3:f64 = 0.1;
+pub const  B1:f32 = 3.0;
+pub const  B2:f32 = 0.60;
+pub const  B3:f32 = 0.1;
 
-pub const  GAMMA1:f64 = 1.0;
-pub const  GAMMA2:f64 = 0.01;
-pub const  GAMMA3:f64 = 1.4;
+pub const  GAMMA1:f32 = 1.0;
+pub const  GAMMA2:f32 = 0.01;
+pub const  GAMMA3:f32 = 1.4;
 
-pub const  DELTA1:f64 = 1.5;
-pub const  DELTA2:f64 = 0.8483;
-pub const  DELTA3:f64 = 16000.0;
-pub const  DELTA4:f64 = 1.25;
-pub const  DELTA5:f64 = 250000.0;
+pub const  DELTA1:f32 = 1.5;
+pub const  DELTA2:f32 = 0.8483;
+pub const  DELTA3:f32 = 16000.0;
+pub const  DELTA4:f32 = 1.25;
+pub const  DELTA5:f32 = 250000.0;
 
-pub const  LAMBDA:f64 = 2.0;
+pub const  LAMBDA:f32 = 2.0;
 
-pub const  QEPSIX2:f64 = 8.0;
-pub const  Q:f64 = 2442.0;
-pub const  PHI:f64 = 6.667;
+pub const  QEPSIX2:f32 = 8.0;
+pub const  Q:f32 = 2442.0;
+pub const  PHI:f32 = 6.667;
 
-pub const  R1:f64 = 12.119;
-pub const  R2:f64 = 20.77;
-pub const  R3:f64 = 3.2;
+pub const  R1:f32 = 12.119;
+pub const  R2:f32 = 20.77;
+pub const  R3:f32 = 3.2;
 
-pub const  PI:f64 = 3.14159;
-pub const  PIX2:f64 = 6.28318;
+pub const  PI:f32 = 3.14159;
+pub const  PIX2:f32 = 6.28318;
 
-pub const  CLOUDCOVER:f64 = 1.0;
+pub const  CLOUDCOVER:f32 = 1.0;
 
-pub const  SNOW_COVER_THRESHOLD:f64 = 0.001;
+pub const  SNOW_COVER_THRESHOLD:f32 = 0.001;
